abstract the action in a repeat ring
A "repeat" ring is a fixed size ring that is designed to facilitate a
certain UX which is intuitive for the repeat operation. This UX may
also be useful for other things, such as, perhaps, yanking and
pasting. Emacs already provides a ring-based UX in these cases, but
aside from providing a core ring data structure, the UX itself isn't
abstracted in a way that can be extended by others without
reimplementation. It's not even clear whether the various "ring"
UX's in Emacs share a ring implementation, and the kill ring doesn't
appear to use Emacs's ring data structure.

Abstracting the specific operation to be performed in connection with
a ring is one useful step to abstracting the broadly useful UX here,
as it makes the operation a superficial parametrization of the
underlying ring.

The next step would be to also abstract the means of populating ring
elements.