don't have names for rings
We can leave it to the client application to retain a reference to a
ring it cares about, and we don't anticipate a need to lookup a ring
by name. As far as the means by which the ring is populated, if it is
through subscription on pub/sub, then we need not couple that with a
name for the ring, and can simply leave it to the application to
subscribe a ring to an appropriate topic.