use a noop key sequence to tag repetitions
This is an attempt to avoid using global state to facilitate distinct
handling of repeated key sequences when storing them and resetting the
head. Specifically, we want to store such sequences but *not* reset
the head.

This attempt doesn't quite work though. Reason unclear at the moment,
but it appears that at least in some cases, the noop prefix is being
parsed piecemeal. If we can avoid that, we might be able to store the
noop sequence in the ring as a placeholder and marker for repetition.

But there might be a better approach where we can just store the
repetition state on the ring itself.