implement repeat-pop
Undo the most recent change and replay the current (virtual) head
command on the ring. This is intended to be run right after ".", but
there is currently no check in place to ensure this.

Check if the virtual head is already rotated, and if it is (implying
we are cycling through repetitions on the ring), remove the most
recently stored entry so that the imminent repetition overwrites it
and we don't accumulate entries on the ring from such cycling. If the
head isn't already rotated, then that means that we've only repeated
the most recently stored entry, which, by virtue of the ring not
storing duplicates, would not have been duplicated when repeated, and
so we don't need to remove it as part of "repeat pop."

This introduces a `repeating` flag on each repeat ring to identify key
sequences to be stored that are being repeated vs which are fresh.
This avoids the need for a sentinel "noop" and also avoids a global
state variable for the purpose.