extract generic ring UX into a separate virtual-ring package
The "ring with recency awareness + state" is a common UX pattern in
Emacs used in, e.g., the kill ring, keyboard macro ring, and more.
Yet, the implementations in each of these cases appears to be
distinct, and it does not appear that any of them use Emacs's built-in
ring type!

The new package layers a thin "view" on top of this built-in ring data
structure to model stateful rotation without affecting recency, and we
now leverage this generic functionality from the repeat-ring package.