remove containing dynaring of repeat rings
This was originally meant mainly to add useful structure to what was
then a necessary data structure to keep track of repeat rings. That
kind of "registration" is no longer required and is now facilitated by
lower level patterns in mantra and pubsub.

This containing "ring of rings" could be reinstated in the future if a
need for it becomes apparent.