clear the repeating flag even in the event of an error
This could happen on a user interrupt (`C-g`), for instance if the
interrupt is recorded and then replayed.

If we don't handle the error, then subsequent parsing would be blocked
as we typically assume that activity should not be parsed if
repetition is in progress.

If the flag is being used to suspend parsing in this way, then we
typically should not receive the repeated sequence parsed afresh. But
in case we do, it isn't guaranteed that it would be parsed in the same
way as it originally was, as some of the repeated activity may not be
in the form of key sequences (e.g., it could be direct insertion of
text into the buffer). Since we know we are in the midst of repeating,
in these cases, we now simply store the known repeated mantra (doing
which is useful, as freshly entering the mantra makes it the most
recent one on the repeat ring) instead of whatever may have been
parsed that time.