don't remove elements from the ring when repeat-popping
This was causing entries to dwindle upon successive repeat-popping.

Formerly we needed to remove entries as the repeated sequence would be
freshly parsed each time, but at the moment, that no longer happens as
we suspend parsing during repetition.

It may still make sense to pop and reinsert the entry to the head of
the ring, but at least, removing it is no longer the right thing.